fix paddings and organize.
