+++ src/components/InfoList.jsx
@@ -11,8 +11,17 @@ function InfoList() {
   const closeModal = () => setModalOpen(false);
 
   return (
-    <div>
+    <div className='md:w-1/3 w-1/2 lg:w-1/4'>
       <ul>
+      <li className="w-full h-full">
+          <button className='w-full h-full shadow-inner bg-argBlue  tracking-wider text-xl  p-1 md:text-base rounded-md' type="button" onClick={openModal}>Soft Skills</button>
+        </li>
+        <li className="hover:cursor-pointer hover:text-argBlue">
+          <button type="button" onClick={openModal}>Blog</button>
+        </li>
+        <li className="hover:cursor-pointer hover:text-argBlue">
+          <button type="button" onClick={openModal}>Languages</button>
+        </li>
         <li className="hover:cursor-pointer hover:text-argBlue">
           <button type="button" onClick={toggleHobbies}>Hobbies</button>
         </li>
@@ -29,15 +38,7 @@ function InfoList() {
             </li>
           </ul>
         )}
-        <li className="hover:cursor-pointer hover:text-argBlue">
-          <button type="button" onClick={openModal}>Music</button>
-        </li>
-        <li className="hover:cursor-pointer hover:text-argBlue">
-          <button type="button" onClick={openModal}>Music</button>
-        </li>
-        <li className="hover:cursor-pointer hover:text-argBlue">
-          <button type="button" onClick={openModal}>Languages</button>
-        </li>
+        
       </ul>
 
       {modalOpen && <Modal handleClose={closeModal} text="Modal Content" />}
