+++ src/sections/ContactSection.jsx
@@ -1,4 +1,4 @@
-import ContactForm from '../components/ContactForm';
+import ContactForm from '../components/Contact/ContactForm';
 
 function ContactSection() {
   return (
